Add delete feature to notes
In this commit, the applications is enhanced by adding an option to delete individual notes. This gives users more control over their content enabling better user experience. To achieve this, SoftDeletes were included in the Notes migration table and the delete method was added to the Livewire showroom. The 'delete' option was also added to the actions dropdown on the app front-end.