"Reduce padding and add border to note component
The padding in the 'note' component was decreased to make the notes more compact, improving the clarity and readability of the note list. A border was added to the bottom of the note block for better separation between notes. Hence, giving each note a more defined space and improving the overall UI."